baserock_backup: Update backup script
Storyboard now has its own MySQL database (the current version
won't work with the MariaDB database that we were using before, due to
lack of some feature).

Also, the script should NOT update the timestamp files if a backup
actually failed. The actual errors will be in the journal but it's more
convenient to use the timestamps to see if some backups aren't working.

Change-Id: Ifbd541518da18a1d273638d5a432c64f68172cb4