baserock_openid_provider: Update on top of Fedora 23, and without Packer
This means we no longer use Packer for anything, which is good.

The switch from Django 1.7 to Django 1.9 caused some problems in the
openid_provider module. Upstream for that module is here:

    https://bitbucket.org/romke/django_openid_provider/

At the time of writing there was no fix upstream for these issues, it
would be good to submit them. We have other unsubmitted changes against
that upstream in our openid_provider code.

One issue was use of import_module (which is now available from
importlib, doesn't need to be imported from django.utils any more).

Another is use of WSGIRequest.REQUEST, which is deprecated since
Django 1.7 and removed in Django 1.9. We now need to use .POST or
.GET to get that info.

Change-Id: I60793aaf0d84d81b89ff59efbe08240d99b7973f