Add python3-core stratum, equivalent of python2-core
There have been some important Python components not available for
Python 3 users, such as PyYAML. This commit adds a python3-core stratu
that contain everything from python2-core, but for use with Python 3.
A couple of things are already in core.morph, so python2-core is not
quite identical to python3-core.

I've added this to Trove system so that <https://gerrit.baserock.org/1339/>
can be merged. It can be added to other systems as needed.

Change-Id: I9b742b942d64b0bc369596a689f33e6bc4c21935