Fix backup configuration
The frontend's internal IP has changed due to redeployment, so we need
to update the whitelisted IP. Also, Fedora 23 doesn't contain the
Python 2 version of PyYAML which the backup-snapshot script requires.
It would be good to just require Python3 for backup-snapshot, but I
don't have time to fix up all the issues with that right now.

Change-Id: Ica6302ff288255870511f1c19ec61c324dcd4288